Move and refactor classes in preprocess
