+++ src/main/config/configuration.properties
@@ -5,7 +5,7 @@
 # See the accompanying LICENSE file for applicable license.
 #DITA-OT configuration properties
 default.cascade = merge
-temp-file-name-scheme = org.dita.dost.module.GenMapAndTopicListModule$DefaultTempFileScheme
+temp-file-name-scheme = org.dita.dost.module.reader.DefaultTempFileScheme
 #filter-attributes =
 #flag-attributes =
 cli.color = true
