+++ src/main/java/org/dita/dost/module/KeyrefModule.java
@@ -10,7 +10,7 @@ package org.dita.dost.module;
 
 import org.dita.dost.exception.DITAOTException;
 import org.dita.dost.log.DITAOTLogger;
-import org.dita.dost.module.GenMapAndTopicListModule.TempFileNameScheme;
+import org.dita.dost.module.reader.TempFileNameScheme;
 import org.dita.dost.pipeline.AbstractPipelineInput;
 import org.dita.dost.pipeline.AbstractPipelineOutput;
 import org.dita.dost.reader.KeyrefReader;
@@ -34,7 +34,6 @@ import java.net.URI;
 import java.util.*;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.toMap;
 import static org.dita.dost.util.Configuration.configuration;
@@ -98,7 +97,7 @@ final class KeyrefModule extends AbstractPipelineModuleImpl {
                 final String cls = Optional
                         .ofNullable(job.getProperty("temp-file-name-scheme"))
                         .orElse(configuration.get("temp-file-name-scheme"));
-                tempFileNameScheme = (GenMapAndTopicListModule.TempFileNameScheme) Class.forName(cls).newInstance();
+                tempFileNameScheme = (TempFileNameScheme) Class.forName(cls).newInstance();
             } catch (InstantiationException | IllegalAccessException | ClassNotFoundException e) {
                 throw new RuntimeException(e);
             }
