+++ src/main/java/org/dita/dost/util/DelayConrefUtils.java
@@ -30,7 +30,7 @@ import javax.xml.transform.stream.StreamResult;
 import org.dita.dost.exception.DITAOTException;
 import org.dita.dost.log.DITAOTLogger;
 
-import org.dita.dost.module.GenMapAndTopicListModule.TempFileNameScheme;
+import org.dita.dost.module.reader.TempFileNameScheme;
 import org.dita.dost.writer.ExportAnchorsFilter;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
