+++ src/test/resources/configuration.properties
@@ -3,7 +3,7 @@ otrelease = 1.2
 otversion = 1.2.3
 chunk.id-generation-scheme = counter
 default.cascade = nomerge
-temp-file-name-scheme = org.dita.dost.module.GenMapAndTopicListModule$DefaultTempFileScheme
+temp-file-name-scheme = org.dita.dost.module.reader.DefaultTempFileScheme
 
 # Integration
 plugindirs = plugins;demo
